MAINT: move v3 API functions to v3.py
